devtrack-backend/requirements.txt
# ✅ FastAPI core
fastapi==0.115.1
uvicorn==0.23.2
starlette==0.46.1

# ✅ DB and ORM
SQLAlchemy==2.0.41
psycopg2-binary==2.9.9  # PostgreSQL driver

# ✅ Security and JWT
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2

# ✅ Environment management
python-dotenv==1.1.0

# ✅ Email validation for user signup
email-validator==2.1.0.post1

# ✅ Async and support libs
anyio==4.2.0
typing_extensions==4.13.1
pydantic==2.11.3
pydantic_core==2.33.1

# ✅ CORS support
requests==2.31.0
